requirements.txt
fastapi~=0.116.1
uvicorn
pymongo~=4.14.0
python-dotenv~=1.1.1

requests~=2.32.4